+++ tests/2xx/236_has_param.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+. common.sh
+
+assert_func has_param $TEST_FAIL $TEST_EMPTY
+assert_func has_param $TEST_FAIL $TEST_EMPTY BLA